+++ packages/spot-consumer/src/types/index.ts
@@ -1,7 +1,5 @@
-import {
-  PriceLevel as GrpcPriceLevel,
-  CosmosCoin as GrpcCosmosCoin,
-} from '@injectivelabs/exchange-api/injective_spot_markets_rpc_pb'
+import { CosmosCoin as GrpcCosmosCoin } from '@injectivelabs/exchange-api/injective_accounts_rpc_pb'
+import { PriceLevel as GrpcPriceLevel } from '@injectivelabs/exchange-api/injective_spot_markets_rpc_pb'
 
 export * from './spot-market'
 export * from './spot-market-chronos'