+++ packages/spot-consumer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.16](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.15...@injectivelabs/spot-consumer@0.0.16) (2021-04-21)
+
+**Note:** Version bump only for package @injectivelabs/spot-consumer
+
 ## [0.0.15](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.14...@injectivelabs/spot-consumer@0.0.15) (2021-04-20)
 
 ### Bug Fixes